Remove left-click video action in favor of context menu
