+++ src/ui/components/App.css
@@ -22,6 +22,7 @@ h2 {
 #graphicsVideo {
   position: absolute;
   transform-origin: top left;
+  cursor: context-menu;
 }
 
 #graphicsVideo {
