+++ src/ui/components/Video.tsx
@@ -12,12 +12,10 @@ import {
   getVideoUrl,
 } from "ui/actions/app";
 import CommentsOverlay from "ui/components/Comments/VideoComments/index";
-import useVideoCommentTool from "ui/components/useVideoCommentTool";
 import useVideoContextMenu from "ui/components/useVideoContextMenu";
 import { getSelectedPrimaryPanel } from "ui/reducers/layout";
 import { getPlayback, isPlaybackStalled } from "ui/reducers/timeline";
 import { useAppSelector } from "ui/setup/hooks";
-import { getRecordingId } from "ui/utils/recording";
 
 import ReplayLogo from "./shared/ReplayLogo";
 import Spinner from "./shared/Spinner";
@@ -34,24 +32,11 @@ export default function Video() {
   const stalled = useAppSelector(isPlaybackStalled);
   const mouseTargetsLoading = useAppSelector(getAreMouseTargetsLoading);
   const highlightedNodesLoading = useAppSelector(getHighlightedNodesLoading);
-  const recordingId = useAppSelector(getRecordingId);
 
   const isPaused = !playback;
   const isNodeTarget = recordingTarget == "node";
 
   const canvasRef = useRef<HTMLCanvasElement>(null);
-
-  const {
-    onClick: onVideoCommentClick,
-    onMouseEnter,
-    onMouseLeave,
-    onMouseMove,
-    tooltip,
-  } = useVideoCommentTool({
-    areMouseTargetsLoading: mouseTargetsLoading,
-    canvasRef,
-    recordingId: recordingId!,
-  });
 
   const { contextMenu, onContextMenu } = useVideoContextMenu({ canvasRef });
 
@@ -87,11 +72,12 @@ export default function Video() {
   };
 
   const onClick = (e: React.MouseEvent) => {
+    // User was trying to select something from the video preview, not add a comment
     if (isNodePickerActive || isNodePickerInitializing) {
       return;
     }
 
-    onVideoCommentClick(e);
+    onContextMenu(e);
   };
 
   const showCommentTool =
@@ -111,9 +97,6 @@ export default function Video() {
         onClick={onClick}
         onContextMenu={onContextMenu}
         onMouseDown={onMouseDown}
-        onMouseEnter={onMouseEnter}
-        onMouseLeave={onMouseLeave}
-        onMouseMove={onMouseMove}
         ref={canvasRef}
       />
       {contextMenu}
@@ -126,7 +109,6 @@ export default function Video() {
           )}
         </CommentsOverlay>
       ) : null}
-      {showCommentTool && tooltip}
       {isNodePickerInitializing ? <Tooltip label="Loading…" targetID="video" /> : null}
       {panel === "cypress" && <CypressToggler />}
       <div id="highlighter-root">
